HW11 30.09.21

+++ src/main/java/Lesson11/Homework/Apple.java
@@ -1,9 +1,9 @@
 package Lesson11.Homework;
 
 public class Apple extends Fruit {
-    private int weight;
+    private float weight;
 
-    public Apple(int weight) {
+    public Apple(float weight) {
         super(weight);
     }
 }
